Add support for message and varint IO
This includes support for reading and writing varints and messages,
but does not include all functions included in the standard
protobuf coded streams classes. Functionality can be extended
as needed in the future.

This functionality is compatible with at least the standard
Java protobuf equivalent functions, but the tests for this
compatibility have been split out into a separate commit as they
add Java code and its related build files, along with other changes
to support this.